+++ pipeline/process_block.go
@@ -26,6 +26,7 @@ func (p *Pipeline) ProcessBlock(block *bstream.Block, obj interface{}) (err erro
 	logger := reqctx.Logger(ctx)
 	defer func() {
 		if r := recover(); r != nil {
+			fmt.Println("error")
 			err = fmt.Errorf("panic at block %s: %s", block, r)
 			logger.Error("panic while process block", zap.Uint64("block_num", block.Num()), zap.Error(err))
 			logger.Error(string(debug.Stack()))