+++ tui2/components/blockselect/blockselect_test.go
@@ -1,12 +1,21 @@
 package blockselect
 
 import (
+	"regexp"
 	"testing"
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/streamingfast/substreams/tui2/common"
 	"github.com/stretchr/testify/assert"
 )
+
+const ansi = "[\u001B\u009B][[\\]()#;?]*(?:(?:(?:[a-zA-Z\\d]*(?:;[a-zA-Z\\d]*)*)?\u0007)|(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PRZcf-ntqry=><~]))"
+
+var re = regexp.MustCompile(ansi)
+
+func Strip(str string) string {
+	return re.ReplaceAllString(str, "")
+}
 
 func TestBlockSelect_Bar(t *testing.T) {
 	b := &BlockSelect{
@@ -16,9 +25,12 @@ func TestBlockSelect_Bar(t *testing.T) {
 		lowBlock:       1,
 		highBlock:      20,
 	}
-	assert.Equal(t, b.View(), `1 --- 20                           
- | | |           |                 
-              18 ^                 `)
+	expected := `┌───────────────────────────────────────────┐
+│1 --- 20                                   │
+│ | | |           |                         │
+│                 ^ Current block: 18       │
+└───────────────────────────────────────────┘`
+	assert.Equal(t, expected, Strip(b.View()))
 }
 
 func TestBlockSelect_Update(t *testing.T) {