+++ wasm/instance.go
@@ -3,7 +3,7 @@ package wasm
 import (
 	"encoding/binary"
 	"fmt"
-
+	"github.com/streamingfast/substreams/manifest"
 	"github.com/streamingfast/substreams/storage/store"
 
 	wasmtime "github.com/bytecodealliance/wasmtime-go/v4"
@@ -96,4 +96,70 @@ func (i *Instance) Cleanup() error {
 	}
 	i.Module.wasmStore.GC()
 	return err
+}
+
+func (i *Instance) IsValidSetStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_SET
+}
+
+func (i *Instance) IsValidSetIfNotExists() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_SET_IF_NOT_EXISTS
+}
+
+func (i *Instance) IsValidAppendStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_APPEND
+}
+
+func (i *Instance) IsValidAddBigIntStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD && i.valueType == manifest.OutputValueTypeBigInt
+}
+
+func (i *Instance) IsValidAddBigDecimalStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD &&
+		(i.valueType == manifest.OutputValueTypeBigDecimal || i.valueType == manifest.OutputValueTypeBigFloat)
+}
+
+func (i *Instance) IsValidAddInt64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD && i.valueType == manifest.OutputValueTypeInt64
+}
+
+func (i *Instance) IsValidAddFloat64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD && i.valueType == manifest.OutputValueTypeFloat64
+}
+
+func (i *Instance) IsValidSetMinInt64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN && i.valueType == manifest.OutputValueTypeInt64
+}
+
+func (i *Instance) IsValidSetMinBigIntStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN && i.valueType == manifest.OutputValueTypeBigInt
+}
+
+func (i *Instance) IsValidSetMinFloat64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN && i.valueType == manifest.OutputValueTypeFloat64
+}
+
+func (i *Instance) IsValidSetMinBigDecimalStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN &&
+		(i.valueType == manifest.OutputValueTypeBigDecimal || i.valueType == manifest.OutputValueTypeBigFloat)
+}
+
+func (i *Instance) IsValidSetMaxInt64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX &&
+		i.valueType == manifest.OutputValueTypeInt64
+}
+
+func (i *Instance) IsValidSetMaxBigIntStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX &&
+		i.valueType == manifest.OutputValueTypeBigInt
+}
+
+func (i *Instance) IsValidSetMaxFloat64Store() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX &&
+		i.valueType == manifest.OutputValueTypeFloat64
+}
+
+func (i *Instance) IsValidSetMaxBigDecimalStore() bool {
+	return i.updatePolicy == pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX &&
+		(i.valueType == manifest.OutputValueTypeBigDecimal || i.valueType == manifest.OutputValueTypeBigFloat)
 }